- Brachacki Beds - Brachacki Bed Animations (Block & Player) - Updated Mixins - Updated Newbery interior - Hums

+++ common/src/main/java/dev/jeryn/audreys_additions/common/blocks/LightBoxBlock.java
@@ -3,17 +3,15 @@ package dev.jeryn.audreys_additions.common.blocks;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.world.item.context.BlockPlaceContext;
-import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.HorizontalDirectionalBlock;
 import net.minecraft.world.level.block.RenderShape;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.StateDefinition;
 import net.minecraft.world.level.block.state.properties.BlockStateProperties;
 import net.minecraft.world.level.block.state.properties.DirectionProperty;
-import net.minecraft.world.level.material.PushReaction;
-import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.phys.shapes.CollisionContext;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -69,11 +67,7 @@ public class LightBoxBlock extends HorizontalDirectionalBlock {
     @Override
     public void onRemove(@NotNull BlockState state, @NotNull Level level, @NotNull BlockPos pos, @NotNull BlockState newState, boolean isMoving) {
         super.onRemove(state, level, pos, newState, isMoving);
-
-        // Determine the position of the secondary block (barrier)
         BlockPos barrierPos = pos.relative(state.getValue(FACING));
-
-        // Remove the barrier block if it exists
         if (level.getBlockState(barrierPos).is(Blocks.BARRIER)) {
             level.removeBlock(barrierPos, false);
         }
@@ -82,13 +76,16 @@ public class LightBoxBlock extends HorizontalDirectionalBlock {
     @Override
     public void neighborChanged(@NotNull BlockState state, @NotNull Level level, @NotNull BlockPos pos, @NotNull Block block, @NotNull BlockPos neighborPos, boolean isMoving) {
         super.neighborChanged(state, level, pos, block, neighborPos, isMoving);
-
-        // If the barrier block is broken, break the main block too
         BlockPos barrierPos = pos.relative(state.getValue(FACING));
-        if (level.getBlockState(barrierPos).isAir()) {
-            level.destroyBlock(pos, true);
+        if (level.getBlockState(pos).is(Blocks.BARRIER)) {
+            if (!level.getBlockState(barrierPos).isAir()) {
+                if (level.getBlockState(barrierPos).isAir()) {
+                    level.destroyBlock(pos, true);
+                }
+            }
         }
     }
+
 
     private boolean canPlaceBlock(Level level, BlockPos pos) {
         // Check if the position is valid for block placement
